Refactor code and move displaymodal to different file
