+++ package.json
@@ -28,6 +28,7 @@
   "homepage": "https://github.com/kaizipaul/webpack-setup#readme",
   "devDependencies": {
     "@babel/core": "^7.19.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
     "@babel/preset-env": "^7.19.4",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^29.2.0",
@@ -37,7 +38,7 @@
     "eslint-plugin-import": "^2.26.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^29.2.0",
+    "jest": "^29.2.1",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
