+++ src/__mocks__/counterComments.js
@@ -1,0 +1,6 @@
+const counterComments = (arrayComments) => {
+  const count = arrayComments.length || 0;
+  return count;
+};
+
+export default counterComments;
